requirements.txt
streamlit
psycopg2-binary
pandas
google-generativeai
python-dotenv
